# openapi/api_config.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger loan application aggregator service - OpenAPI 3.0
 *
 * This is a sample some LoanApplication aggregator
 *
 * API version: 1.0.0
 */

package openapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// ConfigAPIController binds http requests to an api service and writes the service results to the http response
type ConfigAPIController struct {
	service      ConfigAPIServicer
	errorHandler ErrorHandler
}

// ConfigAPIOption for how the controller is set up.
type ConfigAPIOption func(*ConfigAPIController)

// WithConfigAPIErrorHandler inject ErrorHandler into controller
func WithConfigAPIErrorHandler(h ErrorHandler) ConfigAPIOption {
	return func(c *ConfigAPIController) {
		c.errorHandler = h
	}
}

// NewConfigAPIController creates a default api controller
func NewConfigAPIController(s ConfigAPIServicer, opts ...ConfigAPIOption) *ConfigAPIController {
	controller := &ConfigAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the ConfigAPIController
func (c *ConfigAPIController) Routes() Routes {
	return Routes{
		"GetConfigs": Route{
			strings.ToUpper("Get"),
			"/api/v1/configs",
			c.GetConfigs,
		},
		"CreateOrganization": Route{
			strings.ToUpper("Post"),
			"/api/v1/configs",
			c.CreateOrganization,
		},
		"GetConfigById": Route{
			strings.ToUpper("Get"),
			"/api/v1/configs/{id}",
			c.GetConfigById,
		},
		"DeleteConfigById": Route{
			strings.ToUpper("Delete"),
			"/api/v1/configs/{id}",
			c.DeleteConfigById,
		},
		"EditConfigById": Route{
			strings.ToUpper("Patch"),
			"/api/v1/configs/{id}",
			c.EditConfigById,
		},
		"GetConfigByOrganizationId": Route{
			strings.ToUpper("Get"),
			"/api/v1/configs/organizations/{id}",
			c.GetConfigByOrganizationId,
		},
		"DeleteConfigByOrganizationId": Route{
			strings.ToUpper("Delete"),
			"/api/v1/configs/organizations/{id}",
			c.DeleteConfigByOrganizationId,
		},
		"EditConfigByOrganizationId": Route{
			strings.ToUpper("Patch"),
			"/api/v1/configs/organizations/{id}",
			c.EditConfigByOrganizationId,
		},
	}
}

// GetConfigs - get configs
func (c *ConfigAPIController) GetConfigs(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetConfigs(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// CreateOrganization - create new config
func (c *ConfigAPIController) CreateOrganization(w http.ResponseWriter, r *http.Request) {
	var configParam Config
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&configParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertConfigRequired(configParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertConfigConstraints(configParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.CreateOrganization(r.Context(), configParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetConfigById - get config data
func (c *ConfigAPIController) GetConfigById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	idParam, err := parseNumericParameter[int32](
		params["id"],
		WithRequire[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "id", Err: err}, nil)
		return
	}
	result, err := c.service.GetConfigById(r.Context(), idParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteConfigById - Delete config data
func (c *ConfigAPIController) DeleteConfigById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	idParam, err := parseNumericParameter[int32](
		params["id"],
		WithRequire[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "id", Err: err}, nil)
		return
	}
	result, err := c.service.DeleteConfigById(r.Context(), idParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// EditConfigById - update config information
func (c *ConfigAPIController) EditConfigById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	idParam, err := parseNumericParameter[int32](
		params["id"],
		WithRequire[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "id", Err: err}, nil)
		return
	}
	var configParam Config
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&configParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertConfigRequired(configParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertConfigConstraints(configParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.EditConfigById(r.Context(), idParam, configParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetConfigByOrganizationId - get organization's config data
func (c *ConfigAPIController) GetConfigByOrganizationId(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	idParam, err := parseNumericParameter[int32](
		params["id"],
		WithRequire[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "id", Err: err}, nil)
		return
	}
	result, err := c.service.GetConfigByOrganizationId(r.Context(), idParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteConfigByOrganizationId - Delete organization's config data
func (c *ConfigAPIController) DeleteConfigByOrganizationId(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	idParam, err := parseNumericParameter[int32](
		params["id"],
		WithRequire[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "id", Err: err}, nil)
		return
	}
	result, err := c.service.DeleteConfigByOrganizationId(r.Context(), idParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// EditConfigByOrganizationId - update organization's config information
func (c *ConfigAPIController) EditConfigByOrganizationId(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	idParam, err := parseNumericParameter[int32](
		params["id"],
		WithRequire[int32](parseInt32),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "id", Err: err}, nil)
		return
	}
	var configParam Config
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&configParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertConfigRequired(configParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertConfigConstraints(configParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.EditConfigByOrganizationId(r.Context(), idParam, configParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}