# openapi/model_config.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger loan application aggregator service - OpenAPI 3.0
 *
 * This is a sample some LoanApplication aggregator
 *
 * API version: 1.0.0
 */

package openapi




type Config struct {

	OrganizationId int32 `json:"organization_id,omitempty"`

	MaxLoanAmount int32 `json:"max_loan_amount,omitempty"`

	NewClient bool `json:"new_client,omitempty"`

	LoanAvailable bool `json:"loan_available,omitempty"`

	CreatedAt string `json:"created_at,omitempty"`

	UpdatedAt string `json:"updated_at,omitempty"`

	IsAvaliable bool `json:"is_avaliable,omitempty"`
}

// AssertConfigRequired checks if the required fields are not zero-ed
func AssertConfigRequired(obj Config) error {
	return nil
}

// AssertConfigConstraints checks if the values respects the defined constraints
func AssertConfigConstraints(obj Config) error {
	return nil
}